Revert "Origin/hotfix/appeals 46639 release (#22242)" (#22271)
This reverts commit d36f9a9cea8b5979892bd302e029aba792dce50f.
