+++ app/models/appeal.rb
@@ -251,15 +251,11 @@ class Appeal < DecisionReview
 
     category_substrings = %w[Contested Apportionment]
 
-    request_issues.each do |request_issue|
-      category_substrings.each do |substring|
-        if request_issue.active? && request_issue.nonrating_issue_category&.include?(substring)
-          return true
-        end
+    request_issues.active.any? do |request_issue|
+      category_substrings.any? do |substring|
+        request_issues.active.include?(request_issue) && request_issue.nonrating_issue_category&.include?(substring)
       end
     end
-
-    false
   end
 
   # :reek:RepeatedConditionals
