+++ app/models/decision_review.rb
@@ -125,11 +125,7 @@ class DecisionReview < CaseflowRecord
   end
 
   def number_of_issues
-    request_issues.count(&:active?)
-  end
-
-  def issue_categories
-    request_issues.select(&:active?).map(&:nonrating_issue_category)
+    request_issues.active.count
   end
 
   def external_id
