+++ app/models/serializers/work_queue/task_column_serializer.rb
@@ -90,7 +90,11 @@ class WorkQueue::TaskColumnSerializer
     columns = [Constants.QUEUE_CONFIG.COLUMNS.ISSUE_TYPES.name]
 
     if serialize_attribute?(params, columns)
-      object.appeal.issue_categories.join(",")
+      if object.appeal.is_a?(LegacyAppeal)
+        object.appeal.issue_categories
+      else
+        object.appeal.request_issues.active.map(&:nonrating_issue_category)
+      end.join(",")
     end
   end
