--- client/app/queue/caching/cachingReducer.js
@@ -1,8 +1,0 @@
-import { combineReducers } from '@reduxjs/toolkit';
-import queueTableCacheReducer from './queueTableCache.slice';
-
-const cachingReducer = combineReducers({
-  queueTable: queueTableCacheReducer,
-});
-
-export default cachingReducer;
